updated amounts

+++ IndiaEvents.Models/Models/EventTypeSheets/Class2.cs
@@ -32,16 +32,16 @@ namespace IndiaEvents.Models.Models.EventTypeSheets
         public string? VendorName { get; set; }
 
         public string? InitiatorName { get; set; }
-        public int? AdvanceAmount { get; set; }
-        public int? TotalExpenseBTC { get; set; }
-        public int? TotalExpenseBTE { get; set; }
-        public int? TotalHonorariumAmount { get; set; }
-        public int? TotalTravelAccommodationAmount { get; set; }
-        public int? TotalAccomodationAmount { get; set; }
-        public int? TotalBudget { get; set; }
-        public int? TotalLocalConveyance { get; set; }
-        public int? TotalTravelAmount { get; set; }
-        public int? TotalExpense { get; set; }
+        public double? AdvanceAmount { get; set; }
+        public double? TotalExpenseBTC { get; set; }
+        public double? TotalExpenseBTE { get; set; }
+        public double? TotalHonorariumAmount { get; set; }
+        public double? TotalTravelAccommodationAmount { get; set; }
+        public double? TotalAccomodationAmount { get; set; }
+        public double? TotalBudget { get; set; }
+        public double? TotalLocalConveyance { get; set; }
+        public double? TotalTravelAmount { get; set; }
+        public double? TotalExpense { get; set; }
         public string? InitiatorEmail { get; set; }
         public string? RBMorBMEmail { get; set; }
         public string? SalesHeadEmail { get; set; }
@@ -88,29 +88,29 @@ namespace IndiaEvents.Models.Models.EventTypeSheets
         public DateTime? FCPAIssueDate { get; set; }
 
         public string? IsHonorariumApplicable { get; set; }
-        public int? Presentation_Speaking_WorkshopDuration { get; set; }
-        public int? DevelopmentofPresentationPanelSessionPreparation { get; set; }
-        public int? PaneldiscussionSessionduration { get; set; }
-        public int? QASession { get; set; }
-        public int? Speaker_TrainerBriefing { get; set; }
-        public int? TotalNoOfHours { get; set; }
-        public int? HonorariumAmountexcludingTax { get; set; }
-        public int? HonorariumAmountincludingTax { get; set; }
-        public int? YTDspendIncludingCurrentEvent { get; set; }
+        public double? Presentation_Speaking_WorkshopDuration { get; set; }
+        public double? DevelopmentofPresentationPanelSessionPreparation { get; set; }
+        public double? PaneldiscussionSessionduration { get; set; }
+        public double? QASession { get; set; }
+        public double? Speaker_TrainerBriefing { get; set; }
+        public double? TotalNoOfHours { get; set; }
+        public double? HonorariumAmountexcludingTax { get; set; }
+        public double? HonorariumAmountincludingTax { get; set; }
+        public double? YTDspendIncludingCurrentEvent { get; set; }
         public string? ExpenseType { get; set; }
         public string? IsTravelBTC_BTE { get; set; }
         public string? IsAccomodationBTC_BTE { get; set; }
         public string? IsLCBTC_BTE { get; set; }
         public string? TravelSelection { get; set; }
-        public int? TravelAmountExcludingTax { get; set; }
-        public int? TravelAmountIncludingTax { get; set; }
-        public int? AccomodationAmountExcludingTax { get; set; }
-        public int? AccomodationAmountIncludingTax { get; set; }
-        public int? LocalConveyanceAmountexcludingTax { get; set; }
-        public int? LocalConveyanceAmountincludingTax { get; set; }
-        public int? AgreementAmount { get; set; }
-        public int? TravelandAccomodationspendincludingcurrentevent { get; set; }
-        public int? FinalAmount { get; set; }
+        public double? TravelAmountExcludingTax { get; set; }
+        public double? TravelAmountIncludingTax { get; set; }
+        public double? AccomodationAmountExcludingTax { get; set; }
+        public double? AccomodationAmountIncludingTax { get; set; }
+        public double? LocalConveyanceAmountexcludingTax { get; set; }
+        public double? LocalConveyanceAmountincludingTax { get; set; }
+        public double? AgreementAmount { get; set; }
+        public double? TravelandAccomodationspendincludingcurrentevent { get; set; }
+        public double? FinalAmount { get; set; }
         public EventRequestBenificiaryDetails? BenificiaryDetailsData { get; set; }
         public string? IsFilesUpload { get; set; }
         public List<string>? Files { get; set; }
@@ -124,8 +124,8 @@ namespace IndiaEvents.Models.Models.EventTypeSheets
         public string? MISCode { get; set; }
         public string? IsLocalConveyance { get; set; }
         public string? BtcorBte { get; set; }
-        public int? LcAmountIncludingTax { get; set; }
-        public int? LcAmountExcludingTax { get; set; }
+        public double? LcAmountIncludingTax { get; set; }
+        public double? LcAmountExcludingTax { get; set; }
         public string? InviteedFrom { get; set; }
         public string? Invitee_Employee_HcpName { get; set; }
         public string? Speciality { get; set; }
@@ -148,8 +148,8 @@ namespace IndiaEvents.Models.Models.EventTypeSheets
     public class ExpenseSheetData
     {
         public string? ExpenseType { get; set; }
-        public int? ExpenseAmountIncludingTax { get; set; }
-        public int? ExpenseAmountExcludingTax { get; set; }
+        public double? ExpenseAmountIncludingTax { get; set; }
+        public double? ExpenseAmountExcludingTax { get; set; }
         public string? IsBtcorBte { get; set; }
     }
